fix(resources): WB-2615, use distinct registries for old-fashioned and new resource services
